fix counter type remove tab formating

+++ schema.sql
@@ -6,24 +6,24 @@ CREATE DATABASE pastebin OWNER pastebin;
 GRANT ALL PRIVILEGES ON DATABASE pastebin TO pastebin;
 \connect pastebin
 CREATE TABLE pastes(
-	pasteid 	varchar(32) PRIMARY KEY NOT NULL,
-	token 		varchar(32)		NOT NULL,
-	lexer 		varchar(512) 		NOT NULL,
-	expiration 	timestamp 		NOT NULL,
-	burn 		int 			NOT NULL,
-	paste 		text 			NOT NULL,
-	size		int			,
-	lines		int			,
-	sloc		int
+	pasteid varchar(32) PRIMARY KEY NOT NULL,
+	token varchar(32) NOT NULL,
+	lexer varchar(512) NOT NULL,
+	expiration timestamp NOT NULL,
+	burn int NOT NULL,
+	paste text NOT NULL,
+	size int ,
+	lines int ,
+	sloc int
 );
 CREATE TABLE stats (
-	metric	varchar(32) PRIMARY KEY NOT NULL,
-	counter bigserial NOT NULL
+	metric varchar(32) PRIMARY KEY NOT NULL,
+	counter bigint NOT NULL
 );
 CREATE TABLE dailystats (
-	date  		timestamp PRIMARY KEY	NOT NULL,
-	pastecount 	bigserial 		NOT NULL,
-	pasteviews 	bigserial 		NOT NULL
+	date timestamp PRIMARY KEY NOT NULL,
+	pastecount int NOT NULL,
+	pasteviews int NOT NULL
 );
 
 INSERT INTO stats (metric, counter) VALUES ('totalpastes', 0);
